+++ pkg/deploy/ingress.go
@@ -4,6 +4,7 @@ package deploy
 
 import (
 	"fmt"
+	"github.com/vertigobr/safira/pkg/utils"
 	"strconv"
 	"strings"
 
@@ -11,7 +12,7 @@ import (
 )
 
 type ingressSpec struct {
-	Rules []ingressRule  `yaml:"rules,omitempty"`
+	Rules []ingressRule `yaml:"rules,omitempty"`
 }
 
 type ingressRule struct {
@@ -51,13 +52,16 @@ func (k *K8sYaml) MountIngress(ingressName, serviceName, path, hostname, env str
 		return err
 	}
 
-	annotations := GetIngressAnnotations(ingressName, stack.Functions)
+	annotations, err := GetIngressAnnotations(ingressName, stack.Functions)
+	if err != nil {
+		return err
+	}
 
 	*k = K8sYaml{
 		ApiVersion: "extensions/v1beta1",
 		Kind:       "Ingress",
 		Metadata: metadata{
-			Name: ingressName,
+			Name:        ingressName,
 			Annotations: annotations,
 		},
 		Spec: ingressSpec{
@@ -94,26 +98,40 @@ func getGatewayPort(url string) (gateway string, port int, err error) {
 
 	split := strings.Split(gateway, ":")
 	gateway = split[0]
-	port, err = strconv.Atoi(split[1])
-	if err != nil {
-		return "", 0, fmt.Errorf("error ao identificar a porta do hostname: %s", err.Error())
+
+	if len(split) > 1 {
+		port, err = strconv.Atoi(split[1])
+		if err != nil {
+			return "", 0, fmt.Errorf("error ao identificar a porta do hostname: %s", err.Error())
+		}
+	} else {
+		port = 8080
 	}
 
 	return
 }
 
-func GetIngressAnnotations(ingressName string, functions map[string]s.Function) map[string]string {
+func GetIngressAnnotations(ingressName string, functions map[string]s.Function) (map[string]string, error) {
 	annotations := make(map[string]string)
 
 	for functionName, function := range functions {
 		if functionName == ingressName {
 			for pluginName, plugin := range function.Plugins {
 				if plugin.Type == "ingress" {
-					annotations["plugins.konghq.com"] = fmt.Sprintf("%s-%s", functionName, pluginName)
+					annotations["konghq.com/plugins"] = fmt.Sprintf("%s-%s", functionName, pluginName)
 				}
 			}
 		}
 	}
 
-	return annotations
+	repoName, err := utils.GetCurrentFolder()
+	if err != nil {
+		return nil, err
+	}
+
+	annotations["kubernetes.io/ingress.class"] = "kong"
+	annotations["safira.io/repository"] = repoName
+	annotations["safira.io/function"] = ingressName
+
+	return annotations, err
 }