+++ pkg/utils/yaml.go
@@ -30,3 +30,22 @@ func CreateYamlFile(fileName string, bytes []byte, clearFile bool) error {
 
 	return nil
 }
+
+func AppendYamlFile(fileName string, bytes []byte) error {
+	if !strings.HasSuffix(fileName, ".yaml") && !strings.HasSuffix(fileName, ".yml") {
+		fileName = fileName + ".yml"
+	}
+
+	f, err := os.OpenFile(fileName, os.O_WRONLY|os.O_APPEND, 0644)
+	if err != nil {
+		return fmt.Errorf("error ao abrir o arquivo %s: %s", fileName, err.Error())
+	}
+	defer f.Close()
+
+	_, err = f.Write(bytes)
+	if err != nil {
+		return fmt.Errorf("error ao escrever no arquivo %s: %s", fileName, err.Error())
+	}
+
+	return nil
+}